backend/requirements.txt
flask
flask-cors
psutil
scapy
python-nmap